src/agents_metadrive/pyproject.toml
[project]
name = "agents-metadrive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "metadrive",
    "datetime>=6.0",
    "gymnasium>=1.2.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "setuptools<70.0.0",
    "stable-baselines3>=2.7.1",
    "tensorboard>=2.20.0",
    "tqdm>=4.67.3",
]

[tool.uv.sources]
metadrive = { path = "../../metadrive", editable = true }
